pmm, regions: separate memory regions related code from PMM
Add regions.{c,h} files to keep memory regions handling code separated
from PMM functionality. This allows to maintain code better and
isolate PMM from multiboot dependency.
