+++ include/mm/pmm.h
@@ -25,56 +25,10 @@
 #ifndef KTF_PMM_H
 #define KTF_PMM_H
 
-#define BDA_ADDR_START 0x400
-#define BDA_ADDR_END   0x4FF
-
-#define BDA_COM_PORTS_ENTRY 0x400
-#define EBDA_ADDR_ENTRY     0x40E
-
-#define BIOS_ACPI_ROM_START 0xE0000
-#define BIOS_ACPI_ROM_STOP  0xFFFFF
-
-#define BIOS_ROM_ADDR_START 0xF0000
-
 #ifndef __ASSEMBLY__
-#include <cmdline.h>
 #include <list.h>
-#include <page.h>
 
-extern unsigned long __start_text[], __end_text[];
-extern unsigned long __start_data[], __end_data[];
-extern unsigned long __start_bss[], __end_bss[];
-extern unsigned long __start_rodata[], __end_rodata[];
-
-extern unsigned long __start_text_user[], __end_text_user[];
-extern unsigned long __start_data_user[], __end_data_user[];
-extern unsigned long __start_bss_user[], __end_bss_user[];
-
-extern unsigned long __start_text_init[], __end_text_init[];
-extern unsigned long __start_data_init[], __end_data_init[];
-extern unsigned long __start_bss_init[], __end_bss_init[];
-
-extern unsigned long __start_text_rmode[], __end_text_rmode[];
-extern unsigned long __start_data_rmode[], __end_data_rmode[];
-extern unsigned long __start_bss_rmode[], __end_bss_rmode[];
-
-extern struct ktf_param __start_cmdline[], __end_cmdline[];
-
-extern unsigned long __weak __start_symbols[], __end_symbols[];
-
-struct addr_range {
-    const char *name;
-    unsigned long base;
-    unsigned long flags;
-    void *start;
-    void *end;
-};
-typedef struct addr_range addr_range_t;
-
-extern addr_range_t addr_ranges[];
-#define for_each_memory_range(ptr)                                                       \
-    for (addr_range_t *ptr = &addr_ranges[0];                                            \
-         ptr->name != NULL || (ptr->start != 0x0 && ptr->end != 0x0); ptr++)
+#include <mm/regions.h>
 
 struct frame {
     struct list_head list;
@@ -90,15 +44,7 @@ typedef bool (*free_frames_cond_t)(frame_t *free_frame);
 
 /* External definitions */
 
-extern void display_memory_map(void);
 extern void display_frames_count(void);
-
-extern addr_range_t get_memory_range(paddr_t pa);
-extern paddr_t get_memory_range_start(paddr_t pa);
-extern paddr_t get_memory_range_end(paddr_t pa);
-
-extern bool paddr_invalid(paddr_t pa);
-
 extern void init_pmm(void);
 
 extern frame_t *get_free_frames_cond(free_frames_cond_t cb);
@@ -110,37 +56,13 @@ extern void map_used_memory(void);
 
 /* Static definitions */
 
+static inline bool paddr_invalid(paddr_t pa) {
+    return pa == PADDR_INVALID || !has_memory_range(pa);
+}
+
 static inline bool mfn_invalid(mfn_t mfn) { return paddr_invalid(mfn_to_paddr(mfn)); }
 
 static inline mfn_t get_free_frame(void) { return get_free_frames(PAGE_ORDER_4K); }
-
-static inline bool in_text_section(const void *addr) {
-    return (addr >= _ptr(__start_text) && addr < _ptr(__end_text)) ||
-           (addr >= _ptr(__start_text_init) && addr < _ptr(__end_text_init));
-}
-
-static inline bool in_init_section(const void *addr) {
-    return (addr >= _ptr(__start_text_init) && addr < _ptr(__end_text_init)) ||
-           (addr >= _ptr(__start_data_init) && addr < _ptr(__end_data_init)) ||
-           (addr >= _ptr(__start_bss_init) && addr < _ptr(__end_bss_init));
-}
-
-static inline bool in_user_section(const void *addr) {
-    return (addr >= _ptr(__start_text_user) && addr < _ptr(__end_text_user)) ||
-           (addr >= _ptr(__start_data_user) && addr < _ptr(__end_data_user)) ||
-           (addr >= _ptr(__start_bss_user) && addr < _ptr(__end_bss_user));
-}
-
-static inline bool in_kernel_section(const void *addr) {
-    return (addr >= _ptr(__start_text) && addr < _ptr(__end_text)) ||
-           (addr >= _ptr(__start_data) && addr < _ptr(__end_data)) ||
-           (addr >= _ptr(__start_bss) && addr < _ptr(__end_bss)) ||
-           (addr >= _ptr(__start_rodata) && addr < _ptr(__end_rodata));
-}
-
-static inline uint32_t get_bios_ebda_addr(void) {
-    return (*(uint16_t *) paddr_to_virt_kern(EBDA_ADDR_ENTRY)) << 4;
-}
 
 #endif /* __ASSEMBLY__ */
