+++ include/multiboot.h
@@ -290,8 +290,7 @@ struct multiboot_apm_info {
 extern void display_multiboot_mmap(void);
 extern void init_multiboot(multiboot_info_t *mbi, const char **cmdline);
 
-#include <mm/pmm.h>
-#include <page.h>
+#include <mm/regions.h>
 extern void map_multiboot_areas(void);
 extern unsigned mbi_get_avail_memory_ranges_num(void);
 extern int mbi_get_avail_memory_range(unsigned index, addr_range_t *r);
